#625 refactor(shutdown): Be consistent with conventions in common library
