+++ services/shutdown/services/additional/brightness/brightness_service_mock.go
@@ -1,0 +1,13 @@
+package brightness
+
+type MockBrightnessService struct {
+	Brightness int
+}
+
+func (service MockBrightnessService) GetBrightness() int {
+	return service.Brightness
+}
+
+func (service *MockBrightnessService) SetBrightness(value int) {
+	service.Brightness = value
+}
